Revert to using helpers.rb
